requirements.txt
Django>=5.1
psycopg2-binary
djangorestframework
osmnx